+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- 
+- add delete on object and delete_objects on bucket
 - support ruby 3.0
 - [#3](https://github.com/transparentclassroom/fake_s3/pull/3) Dependency updates